@@ -40,59 +40,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "15.2.10",
-    "@angular/common": "15.2.10",
-    "@angular/compiler": "15.2.10",
-    "@angular/core": "15.2.10",
-    "@angular/forms": "15.2.10",
-    "@angular/platform-browser": "15.2.10",
-    "@angular/platform-browser-dynamic": "15.2.10",
-    "@angular/router": "15.2.10",
-    "@delon/abc": "15.2.1",
-    "@delon/acl": "15.2.1",
-    "@delon/auth": "15.2.1",
-    "@delon/cache": "15.2.1",
-    "@delon/form": "15.2.1",
-    "@delon/mock": "15.2.1",
-    "@delon/theme": "15.2.1",
-    "@delon/util": "15.2.1",
+    "@angular/animations": "17.3.10",
+    "@angular/common": "17.3.10",
+    "@angular/compiler": "17.3.10",
+    "@angular/core": "17.3.10",
+    "@angular/forms": "17.3.10",
+    "@angular/platform-browser": "17.3.10",
+    "@angular/platform-browser-dynamic": "17.3.10",
+    "@angular/router": "17.3.10",
+    "@ant-design/icons-angular": "19.0.0",
+    "@delon/abc": "17.0.5",
+    "@delon/acl": "17.3.1",
+    "@delon/auth": "17.3.1",
+    "@delon/cache": "17.3.1",
+    "@delon/form": "17.3.1",
+    "@delon/mock": "17.3.1",
+    "@delon/theme": "17.3.1",
+    "@delon/util": "17.3.1",
+    "@kerwin612/ngx-query-builder": "0.6.4",
     "ajv": "8.12.0",
     "ajv-formats": "2.1.1",
-    "angular-tag-cloud-module": "15.0.0",
+    "angular-tag-cloud-module": "17.0.1",
     "echarts": "5.4.3",
     "jquery": "3.7.1",
+    "marked": "15.0.6",
     "monaco-editor": "0.36.1",
-    "ng-zorro-antd": "15.1.0",
-    "ngx-color-picker": "12.0.1",
-    "ngx-echarts": "15.0.3",
-    "ngx-slick-carousel": "15.0.0",
+    "ng-zorro-antd": "17.4.1",
+    "ngx-color-picker": "16.0.0",
+    "ngx-echarts": "17.2.0",
+    "ngx-markdown": "19.0.0",
+    "ngx-slick-carousel": "17.0.0",
     "rxjs": "7.8.1",
     "screenfull": "6.0.2",
     "slick-carousel": "1.8.1",
     "tslib": "2.6.2",
-    "zone.js": "0.11.8"
+    "zone.js": "0.14.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "15.2.10",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "15.2.10",
-    "@angular/compiler-cli": "15.2.10",
-    "@angular/language-service": "15.2.10",
-    "@delon/testing": "15.2.1",
-    "@types/jasmine": "4.0.3",
-    "@types/jasminewd2": "2.0.12",
-    "@types/node": "12.20.55",
-    "@typescript-eslint/eslint-plugin": "5.49.0",
-    "@typescript-eslint/parser": "5.49.0",
-    "eslint": "8.51.0",
+    "@angular-devkit/build-angular": "17.3.8",
+    "@angular-eslint/builder": "17.5.2",
+    "@angular-eslint/eslint-plugin": "17.5.2",
+    "@angular-eslint/eslint-plugin-template": "17.5.2",
+    "@angular-eslint/schematics": "17.5.2",
+    "@angular-eslint/template-parser": "17.5.2",
+    "@angular/cli": "17.3.8",
+    "@angular/compiler-cli": "17.3.10",
+    "@angular/language-service": "17.3.10",
+    "@delon/testing": "17.3.1",
+    "@types/jasmine": "5.1.4",
+    "@types/jasminewd2": "2.0.13",
+    "@types/node": "20.12.11",
+    "@typescript-eslint/eslint-plugin": "7.8.0",
+    "@typescript-eslint/parser": "7.8.0",
+    "eslint": "8.56.0",
     "eslint-config-prettier": "8.6.0",
-    "eslint-plugin-deprecation": "1.3.3",
+    "eslint-plugin-deprecation": "^3.0.0",
     "eslint-plugin-import": "2.26.0",
-    "eslint-plugin-jsdoc": "39.7.5",
+    "eslint-plugin-jsdoc": "48.2.5",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "eslint-plugin-prettier": "4.2.1",
     "husky": "7.0.4",
@@ -116,7 +120,7 @@
     "stylelint-declaration-block-no-ignored-properties": "2.7.0",
     "stylelint-order": "5.0.0",
     "ts-node": "10.9.1",
-    "typescript": "4.9.5"
+    "typescript": "5.3.3"
   },
   "lint-staged": {
     "(src)/**/*.{html,ts}": [
