@@ -11,12 +11,7 @@ export const features = [
       <>
         <Translate
           values={{
-            docker: <code>{'docker run -d -p 1157:1157 tancloud/hertzbeat'}</code>,
-            console: (
-              <a href={'https://www.console.tancloud.cn'}>
-                Login Now
-              </a>
-            ),
+            docker: <code>{'docker run -d -p 1157:1157 apache/hertzbeat'}</code>,
             br: <br />,
           }}
         >
@@ -51,7 +46,7 @@ export const features = [
           values={{
             github: (
               <a href={'https://github.com/apache/hertzbeat'}>
-                Github Repo
+                Github
               </a>
             ),
             br: <br />,
@@ -62,32 +57,4 @@ export const features = [
       </>
     ),
   },
-]
-
-
-export const usersLink = [
-  {
-    img: 'skyworth_logo.png',
-    alt: 'skyworth',
-  },
-  {
-    img: 'sc_edu_logo.png',
-    alt: 'sc_edu_logo',
-  },
-  {
-    img: 'cmge_logo.png',
-    alt: 'cmge',
-  },
-  {
-    img: 'cnsodata_logo.svg',
-    alt: 'cnsodata',
-  },
-  {
-    img: 'tancloud_logo.svg',
-    alt: 'tancloud',
-  },
-  {
-    img: 'hibobi_logo.svg',
-    alt: 'hibobi',
-  }
 ]
