@@ -61,13 +61,22 @@ import { CONSTANTS } from '../../../shared/constants';
           {{ 'about.help' | i18n }}
         </div>
         <div style="margin-top: 10px; font-weight: bolder; font-size: medium;">
-          <a [href]="'https://github.com/apache/hertzbeat/releases/tag/' + version" target="_blank"> Apache HertzBeat {{ version }} </a>
+          <a [href]="'https://github.com/apache/hertzbeat/releases/tag/' + version" target="_blank">
+            Apache HertzBeat (incubating) {{ version }}
+          </a>
         </div>
         <div style="margin-top: 10px; font-weight: normal; font-size: small;">
           Copyright &copy; {{ currentYear }}
           <nz-divider nzType="vertical"></nz-divider>
-          <a href="https://hertzbeat.apache.org" target="_blank">Apache HertzBeat</a>
+          <a href="https://hertzbeat.apache.org" target="_blank">Apache HertzBeat (incubating)</a>
         </div>
+        <label
+          style="margin-top: 16px;color:gray;font-size:13px"
+          (ngModelChange)="onNotShowAgainChange($event)"
+          nz-checkbox
+          [(ngModel)]="notShowAgain"
+          >{{ 'about.not-show-next-login' | i18n }}
+        </label>
         <nz-divider></nz-divider>
         <div style="margin-top: 10px; font-weight: bolder">
           <span nz-icon nzType="github"></span>
@@ -87,9 +96,11 @@ import { CONSTANTS } from '../../../shared/constants';
           <nz-divider nzType="vertical"></nz-divider>
           <span nz-icon nzType="cloud-download"></span>
           <a href="https://hertzbeat.apache.org/docs/start/upgrade" target="_blank"> {{ 'about.upgrade' | i18n }} </a>
-          <a style="float: right" href="//github.com/apache/hertzbeat" target="_blank"> {{ 'about.star' | i18n }} </a>
-          <span style="float: right" nz-icon nzType="star" nzTheme="twotone"></span>
-          <nz-divider style="float: right" nzType="vertical"></nz-divider>
+          <div style="float: right">
+            <nz-divider nzType="vertical"></nz-divider>
+            <span nz-icon nzType="star" nzTheme="twotone"></span>
+            <a href="//github.com/apache/hertzbeat" target="_blank"> {{ 'about.star' | i18n }} </a>
+          </div>
         </div>
       </div>
     </nz-modal>
@@ -98,21 +109,27 @@ import { CONSTANTS } from '../../../shared/constants';
 })
 export class HeaderUserComponent {
   isAboutModalVisible = false;
+  notShowAgain = false;
   version = CONSTANTS.VERSION;
   currentYear = new Date().getFullYear();
   get user(): User {
     return this.settings.user;
   }
+  private readonly notShowAgainKey = 'NOT_SHOW_ABOUT_NEXT_LOGIN';
 
   constructor(private settings: SettingsService, private router: Router, private localStorageSvc: LocalStorageService) {
+    this.notShowAgain =
+      this.localStorageSvc.getData(this.notShowAgainKey) !== null
+        ? JSON.parse(<string>this.localStorageSvc.getData(this.notShowAgainKey))
+        : false;
     // @ts-ignore
-    if (router.getCurrentNavigation()?.previousNavigation?.finalUrl.toString() === '/passport/login') {
+    if (router.getCurrentNavigation()?.previousNavigation?.finalUrl.toString() === '/passport/login' && !this.notShowAgain) {
       this.showAndCloseAboutModal();
     }
   }
 
   logout(): void {
-    this.localStorageSvc.clear();
+    this.localStorageSvc.clearAuthorization();
     this.router.navigateByUrl('/passport/login');
   }
 
@@ -127,5 +144,10 @@ export class HeaderUserComponent {
   showAndCloseAboutModal() {
     this.isAboutModalVisible = true;
     setTimeout(() => (this.isAboutModalVisible = false), 20000);
+  }
+
+  onNotShowAgainChange(value: boolean): void {
+    this.notShowAgain = value;
+    this.localStorageSvc.putData(this.notShowAgainKey, JSON.stringify(value));
   }
 }
