@@ -4,7 +4,6 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { GlobalFooterModule } from '@delon/abc/global-footer';
-import { NoticeIconModule } from '@delon/abc/notice-icon';
 import { AlainThemeModule } from '@delon/theme';
 import { LayoutDefaultModule } from '@delon/theme/layout-default';
 import { SettingDrawerModule } from '@delon/theme/setting-drawer';
@@ -20,7 +19,6 @@ import { NzInputModule } from 'ng-zorro-antd/input';
 import { NzSpinModule } from 'ng-zorro-antd/spin';
 
 import { LayoutBasicComponent } from './basic/basic.component';
-import { HeaderClearStorageComponent } from './basic/widgets/clear-storage.component';
 import { HeaderFullScreenComponent } from './basic/widgets/fullscreen.component';
 import { HeaderI18nComponent } from './basic/widgets/i18n.component';
 import { HeaderSearchComponent } from './basic/widgets/search.component';
@@ -34,15 +32,19 @@ const HEADER_COMPONENTS = [
   HeaderSearchComponent,
   HeaderFullScreenComponent,
   HeaderI18nComponent,
-  HeaderClearStorageComponent,
   HeaderUserComponent,
   HeaderNotifyComponent
 ];
 
 import { LayoutPassportComponent } from './passport/passport.component';
+import { NzButtonModule } from 'ng-zorro-antd/button';
 import { NzModalModule } from 'ng-zorro-antd/modal';
 import { NzTagModule } from 'ng-zorro-antd/tag';
 import { NzDividerModule } from 'ng-zorro-antd/divider';
+import { NzListComponent, NzListItemActionComponent, NzListItemComponent, NzListItemMetaComponent } from 'ng-zorro-antd/list';
+import { NzStringTemplateOutletDirective } from 'ng-zorro-antd/core/outlet';
+import { NzTooltipDirective } from 'ng-zorro-antd/tooltip';
+import { NzCheckboxComponent } from 'ng-zorro-antd/checkbox';
 const PASSPORT = [LayoutPassportComponent];
 
 @NgModule({
@@ -54,7 +56,6 @@ const PASSPORT = [LayoutPassportComponent];
     ThemeBtnModule,
     SettingDrawerModule,
     LayoutDefaultModule,
-    NoticeIconModule,
     GlobalFooterModule,
     NzDropDownModule,
     NzInputModule,
@@ -65,9 +66,17 @@ const PASSPORT = [LayoutPassportComponent];
     NzBadgeModule,
     NzAvatarModule,
     NzIconModule,
+    NzButtonModule,
     NzModalModule,
     NzTagModule,
-    NzDividerModule
+    NzDividerModule,
+    NzListComponent,
+    NzListItemComponent,
+    NzListItemMetaComponent,
+    NzStringTemplateOutletDirective,
+    NzListItemActionComponent,
+    NzTooltipDirective,
+    NzCheckboxComponent
   ],
   declarations: [...COMPONENTS, ...HEADER_COMPONENTS, ...PASSPORT],
   exports: [...COMPONENTS, ...PASSPORT]
