@@ -20,13 +20,15 @@
 export class NoticeReceiver {
   id!: number;
   name!: string;
-  // 通知信息方式: 0-手机短信 1-邮箱 2-webhook 3-微信公众号 4-企业微信机器人 5-钉钉机器人 6-飞书机器人
-  // 7-Telegram机器人 8-SlackWebHook 9-Discord机器人 10-企业微信应用消息 11-华为云SMN 12-Server酱 13-Gotify
+  // notification mode: 0-sms 1-email 2-webhook 3-wechat public account 4-work wechat robot 5-Dingding robot 6-Feishu robot
+  // 7-Telegram robot 8-SlackWebHook 9-Discord robot 10-work wechat app message 11-Huawei cloud SMN 12-ServerChan 13-Gotify
   type: number = 1;
   phone!: string;
   email!: string;
   tgBotToken!: string;
   tgUserId!: string;
+  tgMessageThreadId!: string;
+  userId!: string;
   slackWebHookUrl!: string;
   discordChannelId!: string;
   discordBotToken!: string;
@@ -36,6 +38,8 @@ export class NoticeReceiver {
   corpId!: string;
   agentId!: number;
   appSecret!: string;
+  partyId!: string;
+  tagId!: string;
   smnAk!: string;
   smnSk!: string;
   smnProjectId!: string;
