@@ -17,23 +17,24 @@
  * under the License.
  */
 
-import { Component, Inject, OnInit } from '@angular/core';
+import { Component, Inject, OnDestroy, OnInit } from '@angular/core';
 import { I18NService } from '@core';
 import { ALAIN_I18N_TOKEN } from '@delon/theme';
 import { NzModalService } from 'ng-zorro-antd/modal';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
-import { NzTableQueryParams } from 'ng-zorro-antd/table';
 
-import { Alert } from '../../../pojo/Alert';
-import { Tag } from '../../../pojo/Tag';
+import { GroupAlert } from '../../../pojo/GroupAlert';
 import { AlertService } from '../../../service/alert.service';
 
+interface ExtendedGroupAlert extends GroupAlert {
+  isNew?: boolean;
+}
 @Component({
   selector: 'app-alert-center',
   templateUrl: './alert-center.component.html',
-  styles: []
+  styleUrl: './alert-center.component.less'
 })
-export class AlertCenterComponent implements OnInit {
+export class AlertCenterComponent implements OnInit, OnDestroy {
   constructor(
     private notifySvc: NzNotificationService,
     private modal: NzModalService,
@@ -44,80 +45,137 @@ export class AlertCenterComponent implements OnInit {
   pageIndex: number = 1;
   pageSize: number = 8;
   total: number = 0;
-  alerts!: Alert[];
+  groupAlerts: ExtendedGroupAlert[] = [];
   tableLoading: boolean = false;
   checkedAlertIds = new Set<number>();
-  // 搜索过滤相关属性
-  filterStatus: number | undefined;
-  filterPriority: number | undefined;
+  filterStatus!: string;
   filterContent: string | undefined;
+  private eventSource!: EventSource;
 
   ngOnInit(): void {
     this.loadAlertsTable();
+    this.initSSESubscription();
   }
 
-  sync() {
-    this.loadAlertsTable();
+  ngOnDestroy(): void {
+    if (this.eventSource) {
+      this.eventSource.close();
+    }
+  }
+
+  // Initialize SSE subscription for real-time alerts
+  private initSSESubscription(): void {
+    this.eventSource = new EventSource('/api/alert/sse/subscribe');
+    this.eventSource.addEventListener('ALERT_EVENT', (evt: MessageEvent) => {
+      try {
+        const newAlert: GroupAlert = JSON.parse(evt.data);
+        this.updateAlertList(newAlert);
+      } catch (error) {
+        console.error('Error parsing SSE data:', error);
+      }
+    });
+
+    // Handle SSE errors
+    this.eventSource.onerror = error => {
+      console.error('SSE connection error:', error);
+      this.eventSource.close();
+    };
+  }
+
+  private updateAlertList(newAlert: GroupAlert): void {
+    const extendedAlert: ExtendedGroupAlert = {
+      ...newAlert,
+      isNew: true
+    };
+
+    if (!extendedAlert.alerts) {
+      extendedAlert.alerts = [];
+    }
+
+    const matchesFilter = this.checkAlertMatchesFilter(extendedAlert);
+    if (!matchesFilter) {
+      return;
+    }
+
+    const existingIndex = this.groupAlerts.findIndex(a => a.id === extendedAlert.id);
+
+    if (existingIndex === -1) {
+      this.groupAlerts = [extendedAlert, ...this.groupAlerts];
+      this.total += 1;
+
+      setTimeout(() => {
+        const index = this.groupAlerts.findIndex(a => a.id === extendedAlert.id);
+        if (index !== -1) {
+          this.groupAlerts[index].isNew = false;
+          // 触发变更检测
+          this.groupAlerts = [...this.groupAlerts];
+        }
+      }, 1000);
+    } else {
+      this.groupAlerts[existingIndex] = {
+        ...extendedAlert,
+        isNew: true
+      };
+
+      setTimeout(() => {
+        if (this.groupAlerts[existingIndex]) {
+          this.groupAlerts[existingIndex].isNew = false;
+          this.groupAlerts = [...this.groupAlerts];
+        }
+      }, 1000);
+
+      this.groupAlerts = [...this.groupAlerts];
+    }
+  }
+
+  private checkAlertMatchesFilter(alert: ExtendedGroupAlert): boolean {
+    if (this.filterStatus && alert.status !== this.filterStatus) {
+      return false;
+    }
+
+    if (this.filterContent) {
+      const searchContent = this.filterContent.toLowerCase();
+
+      const hasMatchingContent = alert.alerts?.some(singleAlert => singleAlert.content?.toLowerCase().includes(searchContent));
+
+      const hasMatchingLabels = Object.entries(alert.groupLabels || {}).some(
+        ([key, value]) => key.toLowerCase().includes(searchContent) || value.toLowerCase().includes(searchContent)
+      );
+
+      if (!hasMatchingContent && !hasMatchingLabels) {
+        return false;
+      }
+    }
+
+    return true;
   }
 
   loadAlertsTable() {
     this.tableLoading = true;
-    let alertsInit$ = this.alertSvc
-      .loadAlerts(this.filterStatus, this.filterPriority, this.filterContent, this.pageIndex - 1, this.pageSize)
-      .subscribe(
-        message => {
-          this.tableLoading = false;
-          this.checkedAll = false;
-          this.checkedAlertIds.clear();
-          if (message.code === 0) {
-            let page = message.data;
-            this.alerts = page.content;
-            this.pageIndex = page.number + 1;
-            this.total = page.totalElements;
-            if (this.alerts) {
-              this.alerts.forEach(item => {
-                item.tmp = [];
-                if (item.tags != undefined) {
-                  Object.keys(item.tags).forEach(name => {
-                    item.tmp.push({
-                      name: name,
-                      value: item.tags[name]
-                    });
-                  });
-                }
-              });
+    let alertsInit$ = this.alertSvc.loadGroupAlerts(this.filterStatus, this.filterContent, this.pageIndex - 1, this.pageSize).subscribe(
+      message => {
+        if (message.code === 0) {
+          let page = message.data;
+          this.groupAlerts = page.content;
+          this.groupAlerts.forEach(alert => {
+            if (alert.alerts == undefined) {
+              alert.alerts = [];
             }
-          } else {
-            console.warn(message.msg);
-          }
-          alertsInit$.unsubscribe();
-        },
-        error => {
-          this.tableLoading = false;
-          alertsInit$.unsubscribe();
-          console.error(error.msg);
+          });
+          this.pageIndex = page.number + 1;
+          this.total = page.totalElements;
+        } else {
+          console.warn(message.msg);
         }
-      );
-  }
-
-  renderAlertTarget(target: string): string {
-    if (target == undefined || target === '') {
-      return target;
-    }
-    const targets = target.split('.');
-    if (targets.length === 3) {
-      let app = this.i18nSvc.fanyi(`monitor.app.${targets[0]}`);
-      let metrics = this.i18nSvc.fanyi(`monitor.app.${targets[0]}.metrics.${targets[1]}`);
-      let metric = this.i18nSvc.fanyi(`monitor.app.${targets[0]}.metrics.${targets[1]}.metric.${targets[2]}`);
-      let value = app == `monitor.app.${targets[0]}` ? targets[0] : app;
-      value += ` / ${metrics == `monitor.app.${targets[0]}.metrics.${targets[1]}` ? targets[1] : metrics}`;
-      value += ` / ${metric == `monitor.app.${targets[0]}.metrics.${targets[1]}.metric.${targets[2]}` ? targets[2] : metric}`;
-      return value;
-    }
-    if (target === 'availability') {
-      return this.i18nSvc.fanyi('monitor.availability');
-    }
-    return target;
+        this.tableLoading = false;
+        alertsInit$.unsubscribe();
+      },
+      error => {
+        this.tableLoading = false;
+        alertsInit$.unsubscribe();
+        console.error(error.msg);
+      }
+    );
   }
 
   onDeleteAlerts() {
@@ -136,18 +194,6 @@ export class AlertCenterComponent implements OnInit {
     });
   }
 
-  onClearAllAlerts() {
-    this.modal.confirm({
-      nzTitle: this.i18nSvc.fanyi('alert.center.confirm.clear-all'),
-      nzOkText: this.i18nSvc.fanyi('common.button.ok'),
-      nzCancelText: this.i18nSvc.fanyi('common.button.cancel'),
-      nzOkDanger: true,
-      nzOkType: 'primary',
-      nzClosable: false,
-      nzOnOk: () => this.clearAllAlerts()
-    });
-  }
-
   onMarkReadAlerts() {
     if (this.checkedAlertIds == null || this.checkedAlertIds.size === 0) {
       this.notifySvc.warning(this.i18nSvc.fanyi('alert.center.notify.no-mark'), '');
@@ -160,7 +206,7 @@ export class AlertCenterComponent implements OnInit {
       nzOkDanger: true,
       nzOkType: 'primary',
       nzClosable: false,
-      nzOnOk: () => this.updateAlertsStatus(this.checkedAlertIds, 3)
+      nzOnOk: () => this.updateAlertsStatus(this.checkedAlertIds, 'resolved')
     });
   }
   onMarkUnReadAlerts() {
@@ -175,7 +221,7 @@ export class AlertCenterComponent implements OnInit {
       nzOkDanger: true,
       nzOkType: 'primary',
       nzClosable: false,
-      nzOnOk: () => this.updateAlertsStatus(this.checkedAlertIds, 0)
+      nzOnOk: () => this.updateAlertsStatus(this.checkedAlertIds, 'firing')
     });
   }
 
@@ -196,18 +242,18 @@ export class AlertCenterComponent implements OnInit {
   onMarkReadOneAlert(alertId: number) {
     let alerts = new Set<number>();
     alerts.add(alertId);
-    this.updateAlertsStatus(alerts, 3);
+    this.updateAlertsStatus(alerts, 'resolved');
   }
 
   onMarkUnReadOneAlert(alertId: number) {
     let alerts = new Set<number>();
     alerts.add(alertId);
-    this.updateAlertsStatus(alerts, 0);
+    this.updateAlertsStatus(alerts, 'firing');
   }
 
   deleteAlerts(alertIds: Set<number>) {
     this.tableLoading = true;
-    const deleteAlerts$ = this.alertSvc.deleteAlerts(alertIds).subscribe(
+    const deleteAlerts$ = this.alertSvc.deleteGroupAlerts(alertIds).subscribe(
       message => {
         deleteAlerts$.unsubscribe();
         if (message.code === 0) {
@@ -232,30 +278,9 @@ export class AlertCenterComponent implements OnInit {
     this.pageIndex = this.pageIndex > lastPage ? lastPage : this.pageIndex;
   }
 
-  clearAllAlerts() {
+  updateAlertsStatus(alertIds: Set<number>, status: string) {
     this.tableLoading = true;
-    const deleteAlerts$ = this.alertSvc.clearAlerts().subscribe(
-      message => {
-        deleteAlerts$.unsubscribe();
-        if (message.code === 0) {
-          this.notifySvc.success(this.i18nSvc.fanyi('common.notify.clear-success'), '');
-          this.loadAlertsTable();
-        } else {
-          this.tableLoading = false;
-          this.notifySvc.error(this.i18nSvc.fanyi('common.notify.clear-fail'), message.msg);
-        }
-      },
-      error => {
-        this.tableLoading = false;
-        deleteAlerts$.unsubscribe();
-        this.notifySvc.error(this.i18nSvc.fanyi('common.notify.clear-fail'), error.msg);
-      }
-    );
-  }
-
-  updateAlertsStatus(alertIds: Set<number>, status: number) {
-    this.tableLoading = true;
-    const markAlertsStatus$ = this.alertSvc.applyAlertsStatus(alertIds, status).subscribe(
+    const markAlertsStatus$ = this.alertSvc.applyGroupAlertsStatus(alertIds, status).subscribe(
       message => {
         markAlertsStatus$.unsubscribe();
         if (message.code === 0) {
@@ -272,37 +297,5 @@ export class AlertCenterComponent implements OnInit {
         this.notifySvc.error(this.i18nSvc.fanyi('common.notify.mark-fail'), error.msg);
       }
     );
-  }
-
-  // begin: 列表多选分页逻辑
-  checkedAll: boolean = false;
-  onAllChecked(checked: boolean) {
-    if (checked) {
-      this.alerts.forEach(monitor => this.checkedAlertIds.add(monitor.id));
-    } else {
-      this.checkedAlertIds.clear();
-    }
-  }
-  onItemChecked(monitorId: number, checked: boolean) {
-    if (checked) {
-      this.checkedAlertIds.add(monitorId);
-    } else {
-      this.checkedAlertIds.delete(monitorId);
-    }
-  }
-  onTablePageChange(params: NzTableQueryParams) {
-    const { pageSize, pageIndex, sort, filter } = params;
-    this.pageIndex = pageIndex;
-    this.pageSize = pageSize;
-    this.loadAlertsTable();
-  }
-  // end: 列表多选分页逻辑
-
-  sliceTagName(tag: Tag): string {
-    if (tag.tagValue != undefined && tag.tagValue.trim() != '') {
-      return `${tag.name}:${tag.tagValue}`;
-    } else {
-      return tag.name;
-    }
   }
 }
