@@ -21,7 +21,9 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import { AlertCenterComponent } from './alert-center/alert-center.component';
-import { AlertConvergeComponent } from './alert-converge/alert-converge.component';
+import { AlertGroupConvergeComponent } from './alert-group/alert-group-converge.component';
+import { AlertInhibitComponent } from './alert-inhibit/alert-inhibit.component';
+import { AlertIntegrationComponent } from './alert-integration/alert-integration.component';
 import { AlertNoticeComponent } from './alert-notice/alert-notice.component';
 import { AlertSettingComponent } from './alert-setting/alert-setting.component';
 import { AlertSilenceComponent } from './alert-silence/alert-silence.component';
@@ -32,8 +34,9 @@ const routes: Routes = [
   { path: 'setting', component: AlertSettingComponent },
   { path: 'notice', component: AlertNoticeComponent },
   { path: 'silence', component: AlertSilenceComponent },
-  { path: 'converge', component: AlertConvergeComponent },
-  { path: '**', component: AlertCenterComponent }
+  { path: 'group', component: AlertGroupConvergeComponent },
+  { path: 'inhibit', component: AlertInhibitComponent },
+  { path: 'integration/:source', component: AlertIntegrationComponent, data: { titleI18n: 'menu.alert.integration' } }
 ];
 
 @NgModule({
