@@ -17,18 +17,17 @@
  * under the License.
  */
 
-import { Component, Inject, OnInit } from '@angular/core';
+import { Component, Inject, OnInit, ViewChild } from '@angular/core';
+import { NgForm } from '@angular/forms';
 import { I18NService } from '@core';
-import { ALAIN_I18N_TOKEN, SettingsService } from '@delon/theme';
+import { ALAIN_I18N_TOKEN } from '@delon/theme';
 import { NzModalService } from 'ng-zorro-antd/modal';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
 import { NzTableQueryParams } from 'ng-zorro-antd/table';
 import { finalize } from 'rxjs/operators';
 
 import { AlertSilence } from '../../../pojo/AlertSilence';
-import { TagItem } from '../../../pojo/NoticeRule';
 import { AlertSilenceService } from '../../../service/alert-silence.service';
-import { TagService } from '../../../service/tag.service';
 
 @Component({
   selector: 'app-alert-silence',
@@ -40,11 +39,10 @@ export class AlertSilenceComponent implements OnInit {
     private modal: NzModalService,
     private notifySvc: NzNotificationService,
     private alertSilenceService: AlertSilenceService,
-    private settingsSvc: SettingsService,
-    private tagService: TagService,
     @Inject(ALAIN_I18N_TOKEN) private i18nSvc: I18NService
   ) {}
 
+  @ViewChild('ruleForm', { static: false }) ruleForm: NgForm | undefined;
   pageIndex: number = 1;
   pageSize: number = 8;
   total: number = 0;
@@ -173,7 +171,7 @@ export class AlertSilenceComponent implements OnInit {
     this.pageIndex = this.pageIndex > lastPage ? lastPage : this.pageIndex;
   }
 
-  // begin: 列表多选分页逻辑
+  // begin: List multiple choice paging
   checkedAll: boolean = false;
   onAllChecked(checked: boolean) {
     if (checked) {
@@ -190,9 +188,9 @@ export class AlertSilenceComponent implements OnInit {
     }
   }
   /**
-   * 分页回调
+   * Paging callback
    *
-   * @param params 页码信息
+   * @param params page info
    */
   onTablePageChange(params: NzTableQueryParams) {
     const { pageSize, pageIndex, sort, filter } = params;
@@ -200,15 +198,13 @@ export class AlertSilenceComponent implements OnInit {
     this.pageSize = pageSize;
     this.loadAlertSilenceTable();
   }
-  // end: 列表多选逻辑
+  // end: List multiple choice paging
 
-  // start 新增修改告警静默 model
+  // start -- new or update alert silence model
   isManageModalVisible = false;
   isManageModalOkLoading = false;
   isManageModalAdd = true;
   silence: AlertSilence = new AlertSilence();
-  searchTag!: string;
-  tagsOption: any[] = [];
   matchTags: string[] = [];
   silenceDates!: Date[];
   dayCheckOptions = [
@@ -268,19 +264,6 @@ export class AlertSilenceComponent implements OnInit {
             this.isManageModalVisible = true;
             this.isManageModalAdd = false;
             this.matchTags = [];
-            if (this.silence.tags != undefined) {
-              this.silence.tags.forEach(item => {
-                let tag = `${item.name}`;
-                if (item.value != undefined) {
-                  tag = `${tag}:${item.value}`;
-                }
-                this.matchTags.push(tag);
-                this.tagsOption.push({
-                  value: tag,
-                  label: tag
-                });
-              });
-            }
           } else {
             this.notifySvc.error(this.i18nSvc.fanyi('common.notify.edit-fail'), message.msg);
           }
@@ -291,21 +274,14 @@ export class AlertSilenceComponent implements OnInit {
       );
   }
   onManageModalOk() {
-    this.silence.tags = [];
-    this.matchTags.forEach(tag => {
-      let tmp: string[] = tag.split(':');
-      let tagItem = new TagItem();
-      if (tmp.length == 1) {
-        tagItem.name = tmp[0];
-        this.silence.tags.push(tagItem);
-      } else if (tmp.length == 2) {
-        tagItem.name = tmp[0];
-        tagItem.value = tmp[1];
-        this.silence.tags.push(tagItem);
-      }
-    });
-    if (this.silence.priorities != undefined) {
-      this.silence.priorities = this.silence.priorities.filter(item => item != null && item != 9);
+    if (this.ruleForm?.invalid) {
+      Object.values(this.ruleForm.controls).forEach(control => {
+        if (control.invalid) {
+          control.markAsDirty();
+          control.updateValueAndValidity({ onlySelf: true });
+        }
+      });
+      return;
     }
     if (this.silence.type === 0) {
       this.silence.periodStart = this.silenceDates[0];
@@ -365,56 +341,8 @@ export class AlertSilenceComponent implements OnInit {
         );
     }
   }
-
-  onPrioritiesChange() {
-    if (this.silence.priorities != undefined) {
-      let isAll = false;
-      this.silence.priorities.forEach(item => {
-        if (item == 9) {
-          isAll = true;
-        }
-      });
-      if (isAll) {
-        this.silence.priorities = [9, 0, 1, 2];
-      }
-    }
-  }
-
-  loadTagsOption() {
-    let tagsInit$ = this.tagService.loadTags(this.searchTag, undefined, 0, 1000).subscribe(
-      message => {
-        if (message.code === 0) {
-          let page = message.data;
-          this.tagsOption = [];
-          if (page.content != undefined) {
-            page.content.forEach(item => {
-              let tag = `${item.name}`;
-              if (item.tagValue != undefined) {
-                tag = `${tag}:${item.tagValue}`;
-              }
-              this.tagsOption.push({
-                value: tag,
-                label: tag
-              });
-            });
-          }
-        } else {
-          console.warn(message.msg);
-        }
-        tagsInit$.unsubscribe();
-      },
-      error => {
-        tagsInit$.unsubscribe();
-        console.error(error.msg);
-      }
-    );
-  }
-
-  sliceTagName(tag: any): string {
-    if (tag.value != undefined && tag.value.trim() != '') {
-      return `${tag.name}:${tag.value}`;
-    } else {
-      return tag.name;
-    }
+  onFilterChange(): void {
+    this.pageIndex = 1;
+    this.loadAlertSilenceTable();
   }
 }
