@@ -18,6 +18,7 @@
  */
 
 import { NgModule, Type } from '@angular/core';
+import { QueryBuilderModule } from '@kerwin612/ngx-query-builder';
 import { SharedModule } from '@shared';
 import { NzBadgeModule } from 'ng-zorro-antd/badge';
 import { NzBreadCrumbModule } from 'ng-zorro-antd/breadcrumb';
@@ -25,16 +26,23 @@ import { NzCascaderModule } from 'ng-zorro-antd/cascader';
 import { NzCollapseModule } from 'ng-zorro-antd/collapse';
 import { NzDatePickerModule } from 'ng-zorro-antd/date-picker';
 import { NzDividerModule } from 'ng-zorro-antd/divider';
+import { NzEmptyModule } from 'ng-zorro-antd/empty';
 import { NzListModule } from 'ng-zorro-antd/list';
+import { NzPaginationModule } from 'ng-zorro-antd/pagination';
 import { NzRadioModule } from 'ng-zorro-antd/radio';
 import { NzSwitchModule } from 'ng-zorro-antd/switch';
 import { NzTagModule } from 'ng-zorro-antd/tag';
 import { NzTimePickerModule } from 'ng-zorro-antd/time-picker';
 import { NzTransferModule } from 'ng-zorro-antd/transfer';
 import { NzUploadModule } from 'ng-zorro-antd/upload';
+import { MarkdownComponent } from 'ngx-markdown';
 
 import { AlertCenterComponent } from './alert-center/alert-center.component';
-import { AlertConvergeComponent } from './alert-converge/alert-converge.component';
+import { AlertGroupConvergeComponent } from './alert-group/alert-group-converge.component';
+import { AlertInhibitComponent } from './alert-inhibit/alert-inhibit.component';
+import { AlertNoticeReceiverComponent } from './alert-notice/alert-notice-receiver/alert-notice-receiver.component';
+import { AlertNoticeRuleComponent } from './alert-notice/alert-notice-rule/alert-notice-rule.component';
+import { AlertNoticeTemplateComponent } from './alert-notice/alert-notice-template/alert-notice-template.component';
 import { AlertNoticeComponent } from './alert-notice/alert-notice.component';
 import { AlertRoutingModule } from './alert-routing.module';
 import { AlertSettingComponent } from './alert-setting/alert-setting.component';
@@ -45,7 +53,11 @@ const COMPONENTS: Array<Type<void>> = [
   AlertSettingComponent,
   AlertNoticeComponent,
   AlertSilenceComponent,
-  AlertConvergeComponent
+  AlertInhibitComponent,
+  AlertGroupConvergeComponent,
+  AlertNoticeReceiverComponent,
+  AlertNoticeRuleComponent,
+  AlertNoticeTemplateComponent
 ];
 
 @NgModule({
@@ -64,7 +76,11 @@ const COMPONENTS: Array<Type<void>> = [
     NzTimePickerModule,
     NzDatePickerModule,
     NzBadgeModule,
-    NzUploadModule
+    NzUploadModule,
+    QueryBuilderModule,
+    NzPaginationModule,
+    NzEmptyModule,
+    MarkdownComponent
   ],
   declarations: COMPONENTS
 })
