@@ -17,8 +17,9 @@
  * under the License.
  */
 
-import { Component, Input } from '@angular/core';
+import { Component, Input, OnInit } from '@angular/core';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
+import { finalize } from 'rxjs/operators';
 
 import { MonitorService } from '../../../service/monitor.service';
 
@@ -27,52 +28,82 @@ import { MonitorService } from '../../../service/monitor.service';
   templateUrl: './monitor-data-table.component.html',
   styleUrls: ['./monitor-data-table.component.less']
 })
-export class MonitorDataTableComponent {
+export class MonitorDataTableComponent implements OnInit {
   @Input()
   get monitorId(): number {
     return this._monitorId;
   }
   set monitorId(monitorId: number) {
     this._monitorId = monitorId;
-    // 需将monitorId作为输入参数的最后一个  这样在执行loadData时其它入参才有值
-    this.loadData();
+    if (this._monitorId && this.metrics) {
+      // Make sure the monitorId is the last input parameter
+      // So that other input parameters are filled in before loadData is executed
+      this.loadData();
+    }
   }
   private _monitorId!: number;
   @Input()
   app!: string;
   @Input()
+  port!: number | undefined;
+  @Input()
+  monitor!: any;
+  @Input()
   metrics!: string;
+  @Input()
+  height: string = '100%';
 
+  showModal!: boolean;
   time!: any;
   fields!: any[];
   valueRows!: any[];
   rowValues!: any[];
   isTable: boolean = true;
+  scrollY: string = '100%';
+  loading: boolean = false;
 
   constructor(private monitorSvc: MonitorService, private notifySvc: NzNotificationService) {}
 
+  ngOnInit(): void {
+    this.scrollY = `calc(${this.height} - 130px)`;
+  }
+
   loadData() {
-    // 读取实时指标数据
-    let metricData$ = this.monitorSvc.getMonitorMetricsData(this.monitorId, this.metrics).subscribe(
-      message => {
-        metricData$.unsubscribe();
-        if (message.code === 0 && message.data) {
-          this.time = message.data.time;
-          this.fields = message.data.fields;
-          this.valueRows = message.data.valueRows;
-          if (this.valueRows.length == 1) {
-            this.isTable = false;
-            this.rowValues = this.valueRows[0].values;
+    this.loading = true;
+    // Read real-time metrics data
+    let metricData$ = this.monitorSvc
+      .getMonitorMetricsData(this.monitorId, this.metrics)
+      .pipe(finalize(() => (this.loading = false)))
+      .subscribe(
+        message => {
+          metricData$.unsubscribe();
+          if (message.code === 0 && message.data) {
+            this.time = message.data.time;
+            this.fields = message.data.fields;
+            this.valueRows = message.data.valueRows;
+            if (this.valueRows.length == 1) {
+              this.isTable = false;
+              this.rowValues = this.valueRows[0].values;
+            }
+          } else if (message.code !== 0) {
+            this.notifySvc.warning(`${this.metrics}:${message.msg}`, '');
+            console.info(`${this.metrics}:${message.msg}`);
           }
-        } else if (message.code !== 0) {
-          this.notifySvc.warning(`${this.metrics}:${message.msg}`, '');
-          console.info(`${this.metrics}:${message.msg}`);
+        },
+        error => {
+          console.error(error.msg);
+          metricData$.unsubscribe();
         }
-      },
-      error => {
-        console.error(error.msg);
-        metricData$.unsubscribe();
-      }
-    );
+      );
+  }
+
+  getObjectLength(obj: Record<string, string> | undefined): number {
+    if (!obj) return 0;
+    return Object.keys(obj).length;
+  }
+
+  getObjectEntries(obj: Record<string, string> | undefined): Array<[string, string]> {
+    if (!obj) return [];
+    return Object.entries(obj);
   }
 }
