@@ -23,19 +23,20 @@ import { ActivatedRoute, ParamMap, Router } from '@angular/router';
 import { I18NService } from '@core';
 import { ALAIN_I18N_TOKEN, TitleService } from '@delon/theme';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
-import { throwError } from 'rxjs';
+import { NzUploadFile } from 'ng-zorro-antd/upload';
+import { of, throwError } from 'rxjs';
 import { finalize, switchMap } from 'rxjs/operators';
 
 import { Collector } from '../../../pojo/Collector';
+import { GrafanaDashboard } from '../../../pojo/GrafanaDashboard';
 import { Message } from '../../../pojo/Message';
 import { Monitor } from '../../../pojo/Monitor';
 import { Param } from '../../../pojo/Param';
 import { ParamDefine } from '../../../pojo/ParamDefine';
-import { Tag } from '../../../pojo/Tag';
 import { AppDefineService } from '../../../service/app-define.service';
 import { CollectorService } from '../../../service/collector.service';
+import { LabelService } from '../../../service/label.service';
 import { MonitorService } from '../../../service/monitor.service';
-import { TagService } from '../../../service/tag.service';
 
 @Component({
   selector: 'app-monitor-modify',
@@ -50,57 +51,60 @@ export class MonitorEditComponent implements OnInit {
     private router: Router,
     private titleSvc: TitleService,
     private notifySvc: NzNotificationService,
-    private tagSvc: TagService,
     private collectorSvc: CollectorService,
+    private labelSvc: LabelService,
     @Inject(ALAIN_I18N_TOKEN) private i18nSvc: I18NService
   ) {}
 
+  sdDefines: ParamDefine[] = [];
+  sdParams: Param[] = [];
   paramDefines!: ParamDefine[];
   hostName!: string;
   params!: Param[];
   advancedParamDefines!: ParamDefine[];
   advancedParams!: Param[];
-  paramValueMap = new Map<String, Param>();
+  paramValueMap!: Map<String, Param>;
   monitor = new Monitor();
+  grafanaDashboard!: GrafanaDashboard;
   collectors!: Collector[];
   collector: string = '';
-  profileForm: FormGroup = new FormGroup({});
-  detected: boolean = false;
-  passwordVisible: boolean = false;
   isSpinning: boolean = false;
   spinningTip: string = 'Loading...';
+  labelKeys: string[] = [];
+  labelMap: { [key: string]: string[] } = {};
 
   ngOnInit(): void {
     this.route.paramMap
       .pipe(
         switchMap((paramMap: ParamMap) => {
           this.isSpinning = true;
-          this.passwordVisible = false;
           let id = paramMap.get('monitorId');
           this.monitor.id = Number(id);
-          // 查询监控信息
+          // query monitor by id
           return this.monitorSvc.getMonitor(this.monitor.id);
         })
       )
       .pipe(
         switchMap((message: Message<any>) => {
           if (message.code === 0) {
+            let paramValueMap = new Map<String, Param>();
             this.monitor = message.data.monitor;
+            if (this.monitor.scrape == null || this.monitor.scrape == undefined) {
+              this.monitor.scrape = 'static';
+            }
+            this.grafanaDashboard = message.data.grafanaDashboard != undefined ? message.data.grafanaDashboard : new GrafanaDashboard();
             this.collector = message.data.collector == null ? '' : message.data.collector;
             this.titleSvc.setTitleByI18n(`monitor.app.${this.monitor.app}`);
             if (message.data.params != null) {
               message.data.params.forEach((item: Param) => {
-                this.paramValueMap.set(item.field, item);
+                paramValueMap.set(item.field, item);
               });
-            }
-            this.detected = false;
-            if (this.monitor.tags == undefined) {
-              this.monitor.tags = [];
+              this.paramValueMap = paramValueMap;
             }
           } else {
             console.warn(message.msg);
-            this.notifySvc.error(this.i18nSvc.fanyi('monitors.not-found'), message.msg);
-            return throwError(this.i18nSvc.fanyi('monitors.not-found'));
+            this.notifySvc.error(this.i18nSvc.fanyi('monitor.not-found'), message.msg);
+            return throwError(this.i18nSvc.fanyi('monitor.not-found'));
           }
           return this.appDefineSvc.getAppParamsDefine(this.monitor.app);
         })
@@ -108,10 +112,10 @@ export class MonitorEditComponent implements OnInit {
       .pipe(
         switchMap(message => {
           if (message.code === 0) {
-            this.params = [];
-            this.advancedParams = [];
-            this.paramDefines = [];
-            this.advancedParamDefines = [];
+            let params: Param[] = [];
+            let advancedParams: Param[] = [];
+            let paramDefines: ParamDefine[] = [];
+            let advancedParamDefines: ParamDefine[] = [];
             message.data.forEach(define => {
               let param = this.paramValueMap.get(define.field);
               if (param === undefined) {
@@ -149,12 +153,15 @@ export class MonitorEditComponent implements OnInit {
                 }
               }
               define.name = this.i18nSvc.fanyi(`monitor.app.${this.monitor.app}.param.${define.field}`);
+              if (define.placeholder == null && this.i18nSvc.fanyi(`monitor.${define.field}.tip`) != `monitor.${define.field}.tip`) {
+                define.placeholder = this.i18nSvc.fanyi(`monitor.${define.field}.tip`);
+              }
               if (define.hide) {
-                this.advancedParams.push(param);
-                this.advancedParamDefines.push(define);
+                advancedParams.push(param);
+                advancedParamDefines.push(define);
               } else {
-                this.params.push(param);
-                this.paramDefines.push(define);
+                params.push(param);
+                paramDefines.push(define);
               }
               if (
                 define.field == 'host' &&
@@ -164,227 +171,235 @@ export class MonitorEditComponent implements OnInit {
                 this.hostName = define.name;
               }
             });
+            this.params = [...params];
+            this.advancedParams = [...advancedParams];
+            this.paramDefines = [...paramDefines];
+            this.advancedParamDefines = [...advancedParamDefines];
           } else {
             console.warn(message.msg);
           }
           return this.collectorSvc.getCollectors();
         })
       )
+      .pipe(
+        switchMap((message: any) => {
+          if (message.code === 0) {
+            this.collectors = message.data.content?.map((item: { collector: any }) => item.collector);
+          } else {
+            console.warn(message.msg);
+          }
+          if (this.monitor.scrape == 'static') {
+            return of({ code: 0, data: [], msg: '' });
+          } else {
+            return this.appDefineSvc.getAppParamsDefine(this.monitor.scrape);
+          }
+        })
+      )
       .subscribe(
         message => {
           if (message.code === 0) {
-            this.collectors = message.data.content?.map(item => item.collector);
+            if (message.data.length > 0) {
+              let params: Param[] = [];
+              let paramDefines: ParamDefine[] = [];
+              message.data.forEach(define => {
+                let param = this.paramValueMap.get(define.field);
+                if (param === undefined) {
+                  param = new Param();
+                  param.field = define.field;
+                  if (define.type === 'number') {
+                    param.type = 0;
+                  } else if (define.type === 'key-value') {
+                    param.type = 3;
+                  } else if (define.type === 'array') {
+                    param.type = 4;
+                  } else {
+                    param.type = 1;
+                  }
+                  if (define.type === 'boolean') {
+                    param.paramValue = define.defaultValue == 'true';
+                  } else if (param.field === 'host') {
+                    param.paramValue = this.monitor.host;
+                  } else if (define.defaultValue != undefined) {
+                    if (define.type === 'number') {
+                      param.paramValue = Number(define.defaultValue);
+                    } else if (define.type === 'boolean') {
+                      param.paramValue = define.defaultValue.toLowerCase() == 'true';
+                    } else {
+                      param.paramValue = define.defaultValue;
+                    }
+                  }
+                } else {
+                  if (define.type === 'boolean') {
+                    if (param.paramValue != null) {
+                      param.paramValue = param.paramValue.toLowerCase() == 'true';
+                    } else {
+                      param.paramValue = false;
+                    }
+                  }
+                }
+                define.name = this.i18nSvc.fanyi(`monitor.app.${this.monitor.scrape}.param.${define.field}`);
+                if (define.placeholder == null && this.i18nSvc.fanyi(`monitor.${define.field}.tip`) != `monitor.${define.field}.tip`) {
+                  define.placeholder = this.i18nSvc.fanyi(`monitor.${define.field}.tip`);
+                }
+                if (define.hide) {
+                  param.display = false;
+                }
+                params.push(param);
+                paramDefines.push(define);
+              });
+              this.sdParams = [...params];
+              this.sdDefines = [...paramDefines];
+            }
           } else {
             console.warn(message.msg);
           }
           this.isSpinning = false;
         },
         error => {
-          console.error(error);
           this.isSpinning = false;
         }
       );
+    this.loadLabels();
   }
 
-  onParamBooleanChanged(booleanValue: boolean, field: string) {
-    // 对SSL的端口联动处理, 不开启SSL默认80端口，开启SSL默认443
-    if (field === 'ssl') {
-      this.params.forEach(param => {
-        if (param.field === 'port') {
-          if (booleanValue) {
-            param.paramValue = '443';
-          } else {
-            param.paramValue = '80';
+  onScrapeChange(scrapeValue: string) {
+    this.monitor.scrape = scrapeValue;
+    if (this.monitor.scrape !== 'static') {
+      this.isSpinning = true;
+      let queryScrapeDefine$ = this.appDefineSvc
+        .getAppParamsDefine(this.monitor.scrape)
+        .pipe(
+          finalize(() => {
+            queryScrapeDefine$.unsubscribe();
+            this.isSpinning = false;
+          })
+        )
+        .subscribe(message => {
+          if (message.code === 0) {
+            let params: Param[] = [];
+            let paramDefines: ParamDefine[] = [];
+            message.data.forEach(define => {
+              let param = new Param();
+              param.field = define.field;
+              if (define.type === 'number') {
+                param.type = 0;
+              } else if (define.type === 'key-value') {
+                param.type = 3;
+              } else if (define.type === 'array') {
+                param.type = 4;
+              } else {
+                param.type = 1;
+              }
+              if (define.type === 'boolean') {
+                param.paramValue = false;
+              }
+              if (define.defaultValue != undefined) {
+                if (define.type === 'number') {
+                  param.paramValue = Number(define.defaultValue);
+                } else if (define.type === 'boolean') {
+                  param.paramValue = define.defaultValue.toLowerCase() == 'true';
+                } else {
+                  param.paramValue = define.defaultValue;
+                }
+              }
+              define.name = this.i18nSvc.fanyi(`monitor.app.${this.monitor.scrape}.param.${define.field}`);
+              if (define.placeholder == null && this.i18nSvc.fanyi(`monitor.${define.field}.tip`) != `monitor.${define.field}.tip`) {
+                define.placeholder = this.i18nSvc.fanyi(`monitor.${define.field}.tip`);
+              }
+              if (define.hide) {
+                param.display = false;
+              }
+              params.push(param);
+              paramDefines.push(define);
+            });
+            this.sdParams = [...params];
+            this.sdDefines = [...paramDefines];
           }
-        }
-      });
+        });
     }
   }
 
-  onSubmit(formGroup: FormGroup) {
-    if (formGroup.invalid) {
-      Object.values(formGroup.controls).forEach(control => {
-        if (control.invalid) {
-          control.markAsDirty();
-          control.updateValueAndValidity({ onlySelf: true });
-        }
-      });
-      return;
-    }
-    this.monitor.host = this.monitor.host.trim();
-    this.monitor.name = this.monitor.name.trim();
-    // todo 暂时单独设置host属性值
-    this.params.forEach(param => {
-      if (param.field === 'host') {
-        param.paramValue = this.monitor.host;
-      }
-      if (param.paramValue != null && typeof param.paramValue == 'string') {
-        param.paramValue = (param.paramValue as string).trim();
-      }
-    });
-    this.advancedParams.forEach(param => {
-      if (param.paramValue != null && typeof param.paramValue == 'string') {
-        param.paramValue = (param.paramValue as string).trim();
-      }
-    });
+  onSubmit(info: any) {
     let addMonitor = {
-      detected: this.detected,
-      monitor: this.monitor,
-      collector: this.collector,
-      params: this.params.concat(this.advancedParams)
+      monitor: info.monitor,
+      collector: info.collector,
+      params: info.params.concat(info.advancedParams).concat(info.sdParams),
+      grafanaDashboard: info.grafanaDashboard
     };
-    if (this.detected) {
-      this.spinningTip = this.i18nSvc.fanyi('monitors.spinning-tip.detecting');
-    } else {
-      this.spinningTip = 'Loading...';
-    }
+    this.spinningTip = 'Loading...';
     this.isSpinning = true;
     this.monitorSvc.editMonitor(addMonitor).subscribe(
       message => {
         this.isSpinning = false;
         if (message.code === 0) {
-          this.notifySvc.success(this.i18nSvc.fanyi('monitors.edit.success'), '');
-          this.router.navigateByUrl(`/monitors?app=${this.monitor.app}`);
+          this.notifySvc.success(this.i18nSvc.fanyi('monitor.edit.success'), '');
+          this.router.navigateByUrl(`/monitors?app=${info.monitor.app}`);
         } else {
-          this.notifySvc.error(this.i18nSvc.fanyi('monitors.edit.failed'), message.msg);
+          this.notifySvc.error(this.i18nSvc.fanyi('monitor.edit.failed'), message.msg);
         }
       },
       error => {
         this.isSpinning = false;
-        this.notifySvc.error(this.i18nSvc.fanyi('monitors.edit.failed'), error.msg);
+        this.notifySvc.error(this.i18nSvc.fanyi('monitor.edit.failed'), error.msg);
       }
     );
   }
 
-  onDetect(formGroup: FormGroup) {
-    if (formGroup.invalid) {
-      Object.values(formGroup.controls).forEach(control => {
-        if (control.invalid) {
-          control.markAsDirty();
-          control.updateValueAndValidity({ onlySelf: true });
-        }
-      });
-      return;
-    }
-    this.monitor.host = this.monitor.host.trim();
-    this.monitor.name = this.monitor.name.trim();
-    // todo 暂时单独设置host属性值
-    this.params.forEach(param => {
-      if (param.field === 'host') {
-        param.paramValue = this.monitor.host;
-      }
-      if (param.paramValue != null && typeof param.paramValue == 'string') {
-        param.paramValue = (param.paramValue as string).trim();
-      }
-    });
-    this.advancedParams.forEach(param => {
-      if (param.paramValue != null && typeof param.paramValue == 'string') {
-        param.paramValue = (param.paramValue as string).trim();
-      }
-    });
+  onDetect(info: any) {
     let detectMonitor = {
-      detected: this.detected,
-      monitor: this.monitor,
-      collector: this.collector,
-      params: this.params.concat(this.advancedParams)
+      monitor: info.monitor,
+      collector: info.collector,
+      params: info.params.concat(info.advancedParams).concat(info.sdParams)
     };
-    this.spinningTip = this.i18nSvc.fanyi('monitors.spinning-tip.detecting');
+    this.spinningTip = this.i18nSvc.fanyi('monitor.spinning-tip.detecting');
     this.isSpinning = true;
     this.monitorSvc.detectMonitor(detectMonitor).subscribe(
       message => {
         this.isSpinning = false;
         if (message.code === 0) {
-          this.notifySvc.success(this.i18nSvc.fanyi('monitors.detect.success'), '');
+          this.notifySvc.success(this.i18nSvc.fanyi('monitor.detect.success'), '');
         } else {
-          this.notifySvc.error(this.i18nSvc.fanyi('monitors.detect.failed'), message.msg);
+          this.notifySvc.error(this.i18nSvc.fanyi('monitor.detect.failed'), message.msg);
         }
       },
       error => {
         this.isSpinning = false;
-        this.notifySvc.error(this.i18nSvc.fanyi('monitors.detect.failed'), error.msg);
+        this.notifySvc.error(this.i18nSvc.fanyi('monitor.detect.failed'), error.msg);
       }
     );
   }
 
   onCancel() {
-    let app = this.monitor.app;
-    app = app ? app : '';
-    this.router.navigateByUrl(`/monitors?app=${app}`);
+    this.router.navigateByUrl(`/monitors`);
   }
 
-  onRemoveTag(tag: Tag) {
-    if (this.monitor != undefined && this.monitor.tags != undefined) {
-      this.monitor.tags = this.monitor.tags.filter(item => item !== tag);
-    }
-  }
-
-  sliceTagName(tag: Tag): string {
-    if (tag.tagValue != undefined && tag.tagValue.trim() != '') {
-      return `${tag.name}:${tag.tagValue}`;
-    } else {
-      return tag.name;
-    }
-  }
-
-  // start Tag model
-  isManageModalVisible = false;
-  isManageModalOkLoading = false;
-  tagCheckedAll: boolean = false;
-  tagTableLoading = false;
-  tagSearch!: string;
-  tags!: Tag[];
-  checkedTags = new Set<Tag>();
-  loadTagsTable() {
-    this.tagTableLoading = true;
-    let tagsReq$ = this.tagSvc.loadTags(this.tagSearch, 1, 0, 1000).subscribe(
+  loadLabels() {
+    let labelsInit$ = this.labelSvc.loadLabels(undefined, undefined, 0, 9999).subscribe(
       message => {
-        this.tagTableLoading = false;
-        this.tagCheckedAll = false;
-        this.checkedTags.clear();
         if (message.code === 0) {
           let page = message.data;
-          this.tags = page.content;
+          this.labelKeys = [...new Set(page.content.map(label => label.name))];
+
+          this.labelMap = {};
+
+          page.content.forEach(label => {
+            if (!this.labelMap[label.name]) {
+              this.labelMap[label.name] = [];
+            }
+
+            if (label.tagValue && !this.labelMap[label.name].includes(label.tagValue)) {
+              this.labelMap[label.name].push(label.tagValue);
+            }
+          });
         } else {
           console.warn(message.msg);
         }
-        tagsReq$.unsubscribe();
+        labelsInit$.unsubscribe();
       },
       error => {
-        this.tagTableLoading = false;
-        tagsReq$.unsubscribe();
+        labelsInit$.unsubscribe();
+        console.error(error.msg);
       }
     );
   }
-  onShowTagsModal() {
-    this.isManageModalVisible = true;
-    this.loadTagsTable();
-  }
-  onManageModalCancel() {
-    this.isManageModalVisible = false;
-  }
-  onManageModalOk() {
-    this.isManageModalOkLoading = true;
-    this.checkedTags.forEach(item => {
-      if (this.monitor.tags.find(tag => tag.id == item.id) == undefined) {
-        this.monitor.tags.push(item);
-      }
-    });
-    this.isManageModalOkLoading = false;
-    this.isManageModalVisible = false;
-  }
-  onAllChecked(checked: boolean) {
-    if (checked) {
-      this.tags.forEach(tag => this.checkedTags.add(tag));
-    } else {
-      this.checkedTags.clear();
-    }
-  }
-  onItemChecked(tag: Tag, checked: boolean) {
-    if (checked) {
-      this.checkedTags.add(tag);
-    } else {
-      this.checkedTags.delete(tag);
-    }
-  }
-  // end tag model
 }
