@@ -22,15 +22,17 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { CollectorComponent } from './collector/collector.component';
 import { DefineComponent } from './define/define.component';
+import { SettingLabelComponent } from './label/label.component';
+import { SettingPluginsComponent } from './plugins/plugin.component';
 import { MessageServerComponent } from './settings/message-server/message-server.component';
 import { ObjectStoreComponent } from './settings/object-store/object-store.component';
 import { SettingsComponent } from './settings/settings.component';
 import { SystemConfigComponent } from './settings/system-config/system-config.component';
 import { StatusComponent } from './status/status.component';
-import { SettingTagsComponent } from './tags/tags.component';
 
 const routes: Routes = [
-  { path: 'tags', component: SettingTagsComponent },
+  { path: 'labels', component: SettingLabelComponent },
+  { path: 'plugins', component: SettingPluginsComponent },
   { path: 'collector', component: CollectorComponent },
   { path: 'status', component: StatusComponent },
   { path: 'define', component: DefineComponent, data: { titleI18n: 'menu.advanced.define' } },
