@@ -26,32 +26,37 @@ import { NzCascaderModule } from 'ng-zorro-antd/cascader';
 import { NzCodeEditorModule } from 'ng-zorro-antd/code-editor';
 import { NzCollapseModule } from 'ng-zorro-antd/collapse';
 import { NzDividerModule } from 'ng-zorro-antd/divider';
+import { NzDropDownModule } from 'ng-zorro-antd/dropdown';
 import { NzLayoutModule } from 'ng-zorro-antd/layout';
 import { NzListModule } from 'ng-zorro-antd/list';
+import { NzPaginationComponent } from 'ng-zorro-antd/pagination';
 import { NzRadioModule } from 'ng-zorro-antd/radio';
 import { NzSwitchModule } from 'ng-zorro-antd/switch';
 import { NzTagModule } from 'ng-zorro-antd/tag';
+import { NzUploadComponent } from 'ng-zorro-antd/upload';
 import { ColorPickerModule } from 'ngx-color-picker';
 
 import { CollectorComponent } from './collector/collector.component';
 import { DefineComponent } from './define/define.component';
+import { SettingLabelComponent } from './label/label.component';
+import { SettingPluginsComponent } from './plugins/plugin.component';
 import { SettingRoutingModule } from './setting-routing.module';
 import { MessageServerComponent } from './settings/message-server/message-server.component';
 import { ObjectStoreComponent } from './settings/object-store/object-store.component';
 import { SettingsComponent } from './settings/settings.component';
 import { SystemConfigComponent } from './settings/system-config/system-config.component';
 import { StatusComponent } from './status/status.component';
-import { SettingTagsComponent } from './tags/tags.component';
 
 const COMPONENTS: Array<Type<void>> = [
-  SettingTagsComponent,
+  SettingLabelComponent,
   DefineComponent,
   SettingsComponent,
   MessageServerComponent,
   SystemConfigComponent,
   ObjectStoreComponent,
   CollectorComponent,
-  StatusComponent
+  StatusComponent,
+  SettingPluginsComponent
 ];
 
 @NgModule({
@@ -70,7 +75,10 @@ const COMPONENTS: Array<Type<void>> = [
     NzCodeEditorModule,
     ClipboardModule,
     NzBadgeModule,
-    NzRadioModule
+    NzRadioModule,
+    NzUploadComponent,
+    NzPaginationComponent,
+    NzDropDownModule
   ],
   declarations: COMPONENTS
 })
