@@ -17,8 +17,8 @@
  * under the License.
  */
 
-import { Component, Inject, OnInit } from '@angular/core';
-import { FormGroup } from '@angular/forms';
+import { Component, Inject, OnInit, ViewChild } from '@angular/core';
+import { FormGroup, NgForm } from '@angular/forms';
 import { I18NService } from '@core';
 import { ALAIN_I18N_TOKEN } from '@delon/theme';
 import { NzModalService } from 'ng-zorro-antd/modal';
@@ -26,13 +26,11 @@ import { NzNotificationService } from 'ng-zorro-antd/notification';
 import { switchMap } from 'rxjs';
 
 import { Message } from '../../../pojo/Message';
-import { TagItem } from '../../../pojo/NoticeRule';
 import { StatusPageComponent } from '../../../pojo/StatusPageComponent';
 import { StatusPageIncident } from '../../../pojo/StatusPageIncident';
 import { StatusPageIncidentContent } from '../../../pojo/StatusPageIncidentContent';
 import { StatusPageOrg } from '../../../pojo/StatusPageOrg';
 import { StatusPageService } from '../../../service/status-page.service';
-import { TagService } from '../../../service/tag.service';
 
 @Component({
   selector: 'app-status',
@@ -44,11 +42,13 @@ export class StatusComponent implements OnInit {
     private notifySvc: NzNotificationService,
     private modal: NzModalService,
     private statusPageService: StatusPageService,
-    private tagService: TagService,
     @Inject(ALAIN_I18N_TOKEN) private i18nSvc: I18NService
   ) {}
 
+  @ViewChild('incidentForm') incidentForm!: NgForm;
+  @ViewChild('componentForm') componentForm!: NgForm;
   statusOrg: StatusPageOrg = new StatusPageOrg();
+  statusOrgForEdit: StatusPageOrg = new StatusPageOrg();
   statusComponents!: StatusPageComponent[];
   statusIncidences!: StatusPageIncident[];
   loading: boolean = false;
@@ -67,8 +67,6 @@ export class StatusComponent implements OnInit {
   isIncidentModalAdd: boolean = true;
 
   search!: string;
-  tagsOption: any[] = [];
-  matchTag: string = '';
 
   ngOnInit(): void {
     this.loadStatusPageConfig();
@@ -80,25 +78,6 @@ export class StatusComponent implements OnInit {
 
   syncIncidence() {
     this.loadIncidenceInfo();
-  }
-
-  loadOrgInfo() {
-    this.orgLoading = true;
-    let orgLoad$ = this.statusPageService.getStatusPageOrg().subscribe(
-      message => {
-        if (message.code === 0) {
-          this.statusOrg = message.data;
-        } else {
-          console.log(message.msg);
-        }
-        this.orgLoading = false;
-        orgLoad$.unsubscribe();
-      },
-      error => {
-        this.orgLoading = false;
-        orgLoad$.unsubscribe();
-      }
-    );
   }
 
   loadComponentInfo() {
@@ -151,6 +130,7 @@ export class StatusComponent implements OnInit {
             this.statusOrg = new StatusPageOrg();
             console.log(message.msg);
           }
+          this.statusOrgForEdit = { ...this.statusOrg };
           return this.statusPageService.getStatusPageComponents();
         })
       )
@@ -177,7 +157,7 @@ export class StatusComponent implements OnInit {
       });
       return;
     }
-    let saveStatus$ = this.statusPageService.saveStatusPageOrg(this.statusOrg).subscribe(
+    let saveStatus$ = this.statusPageService.saveStatusPageOrg(this.statusOrgForEdit).subscribe(
       (message: Message<StatusPageOrg>) => {
         if (message.code === 0) {
           this.statusOrg = message.data;
@@ -197,7 +177,6 @@ export class StatusComponent implements OnInit {
   onNewStatusComponent() {
     this.isComponentModalAdd = true;
     this.currentStatusComponent = new StatusPageComponent();
-    this.matchTag = '';
     this.currentComponentVisible = true;
   }
 
@@ -234,14 +213,7 @@ export class StatusComponent implements OnInit {
 
   onEditOneComponent(data: StatusPageComponent) {
     this.isComponentModalAdd = false;
-    this.currentStatusComponent = data;
-    if (this.currentStatusComponent.tag != undefined) {
-      this.matchTag = this.sliceTagName(this.currentStatusComponent.tag);
-      this.tagsOption.push({
-        value: this.matchTag,
-        label: this.matchTag
-      });
-    }
+    this.currentStatusComponent = { ...data };
     this.currentComponentVisible = true;
   }
 
@@ -310,16 +282,14 @@ export class StatusComponent implements OnInit {
   }
 
   onComponentModalOk() {
-    if (this.matchTag != undefined && this.matchTag.trim() != '') {
-      let tmp: string[] = this.matchTag.split(':');
-      let tagItem = new TagItem();
-      if (tmp.length == 1) {
-        tagItem.name = tmp[0];
-      } else if (tmp.length == 2) {
-        tagItem.name = tmp[0];
-        tagItem.value = tmp[1];
-      }
-      this.currentStatusComponent.tag = tagItem;
+    if (this.componentForm.invalid) {
+      Object.values(this.componentForm.controls).forEach(control => {
+        if (control.invalid) {
+          control.markAsDirty();
+          control.updateValueAndValidity({ onlySelf: true });
+        }
+      });
+      return;
     }
     if (this.statusOrg.id == undefined) {
       this.notifySvc.warning(this.i18nSvc.fanyi('status.component.notify.need-org'), '');
@@ -362,6 +332,15 @@ export class StatusComponent implements OnInit {
   onIncidentModalOk() {
     if (this.statusOrg.id == undefined) {
       this.notifySvc.warning(this.i18nSvc.fanyi('status.component.notify.need-org'), '');
+      return;
+    }
+    if (this.incidentForm.invalid) {
+      Object.values(this.incidentForm.controls).forEach(control => {
+        if (control.invalid) {
+          control.markAsDirty();
+          control.updateValueAndValidity({ onlySelf: true });
+        }
+      });
       return;
     }
     // incident message content
@@ -476,36 +455,6 @@ export class StatusComponent implements OnInit {
     });
   }
 
-  loadTagsOption() {
-    let tagsInit$ = this.tagService.loadTags(undefined, undefined, 0, 1000).subscribe(
-      message => {
-        if (message.code === 0) {
-          let page = message.data;
-          this.tagsOption = [];
-          if (page.content != undefined) {
-            page.content.forEach(item => {
-              let tag = `${item.name}`;
-              if (item.tagValue != undefined) {
-                tag = `${tag}:${item.tagValue}`;
-              }
-              this.tagsOption.push({
-                value: tag,
-                label: tag
-              });
-            });
-          }
-        } else {
-          console.warn(message.msg);
-        }
-        tagsInit$.unsubscribe();
-      },
-      error => {
-        tagsInit$.unsubscribe();
-        console.error(error.msg);
-      }
-    );
-  }
-
   getLatestIncidentContentMsg(incidents: StatusPageIncidentContent[]): string {
     if (incidents == undefined || incidents.length == 0) {
       return '';
@@ -530,14 +479,19 @@ export class StatusComponent implements OnInit {
     }
   }
 
-  sliceTagName(tag: TagItem): string {
-    if (tag == undefined) {
-      return '';
+  getLabelColor(key: string): string {
+    const colors = ['blue', 'green', 'orange', 'purple', 'cyan'];
+    const index = Math.abs(this.hashString(key)) % colors.length;
+    return colors[index];
+  }
+
+  private hashString(str: string): number {
+    let hash = 0;
+    for (let i = 0; i < str.length; i++) {
+      const char = str.charCodeAt(i);
+      hash = (hash << 5) - hash + char;
+      hash = hash & hash;
     }
-    if (tag.value != undefined && tag.value.trim() != '') {
-      return `${tag.name}:${tag.value}`;
-    } else {
-      return tag.name;
-    }
+    return hash;
   }
 }
