@@ -59,18 +59,18 @@ export class AlertDefineService {
   public deleteAlertDefines(alertDefineIds: Set<number>): Observable<Message<any>> {
     let httpParams = new HttpParams();
     alertDefineIds.forEach(alertDefineId => {
-      // 注意HttpParams是不可变对象 需要保存append后返回的对象为最新对象
-      // append方法可以叠加同一key, set方法会把key之前的值覆盖只留一个key-value
+      // HttpParams is unmodifiable, so we need to save the return value of append/set
+      // Method append can append same key, set will replace the previous value
       httpParams = httpParams.append('ids', alertDefineId);
     });
     const options = { params: httpParams };
     return this.http.delete<Message<any>>(alert_defines_uri, options);
   }
 
-  public getAlertDefines(search: string | undefined, pageIndex: number, pageSize: number): Observable<Message<Page<AlertDefine>>> {
+  public getAlertDefines(search: string[] | undefined, pageIndex: number, pageSize: number): Observable<Message<Page<AlertDefine>>> {
     pageIndex = pageIndex ? pageIndex : 0;
     pageSize = pageSize ? pageSize : 8;
-    // 注意HttpParams是不可变对象 需要保存set后返回的对象为最新对象
+    // HttpParams is unmodifiable, so we need to save the return value of append/set
     let httpParams = new HttpParams();
     httpParams = httpParams.appendAll({
       sort: 'id',
@@ -78,8 +78,9 @@ export class AlertDefineService {
       pageIndex: pageIndex,
       pageSize: pageSize
     });
-    if (search != undefined && search.trim() != '') {
-      httpParams = httpParams.append('search', search.trim());
+    if (search != undefined && search.length > 0) {
+      const searchJson = JSON.stringify(search);
+      httpParams = httpParams.append('search', encodeURIComponent(searchJson));
     }
     const options = { params: httpParams };
     return this.http.get<Message<Page<AlertDefine>>>(alert_defines_uri, options);
@@ -88,8 +89,8 @@ export class AlertDefineService {
   public exportAlertDefines(defineIds: Set<number>, type: string): Observable<HttpResponse<Blob>> {
     let httpParams = new HttpParams();
     defineIds.forEach(defineId => {
-      // 注意HttpParams是不可变对象 需要保存append后返回的对象为最新对象
-      // append方法可以叠加同一key, set方法会把key之前的值覆盖只留一个key-value
+      // HttpParams is unmodifiable, so we need to save the return value of append/set
+      // Method append can append same key, set will replace the previous value
       httpParams = httpParams.append('ids', defineId);
     });
     httpParams = httpParams.append('type', type);
