@@ -21,14 +21,15 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { Alert } from '../pojo/Alert';
+import { GroupAlert } from '../pojo/GroupAlert';
 import { Message } from '../pojo/Message';
 import { Page } from '../pojo/Page';
+import { SingleAlert } from '../pojo/SingleAlert';
 
-const alerts_uri = '/alerts';
-const alerts_clear_uri = '/alerts/clear';
 const alerts_summary_uri = '/alerts/summary';
-const alerts_status_uri = '/alerts/status';
+const alerts_group_uri = '/alerts/group';
+const alerts_group_status_uri = '/alerts/group/status';
+const alerts_uri = '/alerts';
 
 @Injectable({
   providedIn: 'root'
@@ -37,59 +38,77 @@ export class AlertService {
   constructor(private http: HttpClient) {}
 
   public loadAlerts(
-    status: number | undefined,
-    priority: number | undefined,
-    content: string | undefined,
+    status: string | undefined,
+    search: string | undefined,
     pageIndex: number,
     pageSize: number
-  ): Observable<Message<Page<Alert>>> {
+  ): Observable<Message<Page<SingleAlert>>> {
     pageIndex = pageIndex ? pageIndex : 0;
     pageSize = pageSize ? pageSize : 8;
-    // 注意HttpParams是不可变对象 需要保存set后返回的对象为最新对象
+    // HttpParams is unmodifiable, so we need to save the return value of append/set
     let httpParams = new HttpParams();
     httpParams = httpParams.appendAll({
-      sort: 'id',
+      sort: 'gmtUpdate',
       order: 'desc',
       pageIndex: pageIndex,
       pageSize: pageSize
     });
-    if (status != undefined && status != 9) {
+    if (status != undefined) {
       httpParams = httpParams.append('status', status);
     }
-    if (priority != undefined && priority != 9) {
-      httpParams = httpParams.append('priority', priority);
-    }
-    if (content != undefined && content != '' && content.trim() != '') {
-      httpParams = httpParams.append('content', content.trim());
+    if (search != undefined && search != '' && search.trim() != '') {
+      httpParams = httpParams.append('search', search.trim());
     }
     const options = { params: httpParams };
-    return this.http.get<Message<Page<Alert>>>(alerts_uri, options);
+    return this.http.get<Message<Page<SingleAlert>>>(alerts_uri, options);
   }
 
-  public deleteAlerts(alertIds: Set<number>): Observable<Message<any>> {
+  public loadGroupAlerts(
+    status: string | undefined,
+    search: string | undefined,
+    pageIndex: number,
+    pageSize: number
+  ): Observable<Message<Page<GroupAlert>>> {
+    pageIndex = pageIndex ? pageIndex : 0;
+    pageSize = pageSize ? pageSize : 8;
+    // HttpParams is unmodifiable, so we need to save the return value of append/set
+    let httpParams = new HttpParams();
+    httpParams = httpParams.appendAll({
+      sort: 'gmtUpdate',
+      order: 'desc',
+      pageIndex: pageIndex,
+      pageSize: pageSize
+    });
+    if (status != undefined) {
+      httpParams = httpParams.append('status', status);
+    }
+    if (search != undefined && search != '' && search.trim() != '') {
+      httpParams = httpParams.append('search', search.trim());
+    }
+    const options = { params: httpParams };
+    return this.http.get<Message<Page<GroupAlert>>>(alerts_group_uri, options);
+  }
+
+  public deleteGroupAlerts(alertIds: Set<number>): Observable<Message<any>> {
     let httpParams = new HttpParams();
     alertIds.forEach(alertId => {
-      // 注意HttpParams是不可变对象 需要保存append后返回的对象为最新对象
-      // append方法可以叠加同一key, set方法会把key之前的值覆盖只留一个key-value
+      // HttpParams is unmodifiable, so we need to save the return value of append/set
+      // Method append can append same key, set will replace the previous value
       httpParams = httpParams.append('ids', alertId);
     });
     const options = { params: httpParams };
-    return this.http.delete<Message<any>>(alerts_uri, options);
+    return this.http.delete<Message<any>>(alerts_group_uri, options);
   }
 
-  public clearAlerts(): Observable<Message<any>> {
-    return this.http.delete<Message<any>>(alerts_clear_uri);
-  }
-
-  public applyAlertsStatus(alertIds: Set<number>, status: number): Observable<Message<any>> {
+  public applyGroupAlertsStatus(alertIds: Set<number>, status: string): Observable<Message<any>> {
     let httpParams = new HttpParams();
     alertIds.forEach(alertId => {
-      // 注意HttpParams是不可变对象 需要保存append后返回的对象为最新对象
-      // append方法可以叠加同一key, set方法会把key之前的值覆盖只留一个key-value
+      // HttpParams is unmodifiable, so we need to save the return value of append/set
+      // Method append can append same key, set will replace the previous value
       httpParams = httpParams.append('ids', alertId);
     });
     const options = { params: httpParams };
-    return this.http.put<Message<any>>(`${alerts_status_uri}/${status}`, null, options);
+    return this.http.put<Message<any>>(`${alerts_group_status_uri}/${status}`, null, options);
   }
 
   public getAlertsSummary(): Observable<Message<any>> {
