@@ -18,5 +18,5 @@
  */
 
 export const CONSTANTS = {
-  VERSION: 'v1.5.0'
+  VERSION: 'v1.7.1'
 };
