@@ -2,15 +2,32 @@ import { CommonModule } from '@angular/common';
 import { NgModule, Type } from '@angular/core';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
+import { IconDefinition } from '@ant-design/icons-angular';
+import { RobotOutline, CloseOutline, SendOutline } from '@ant-design/icons-angular/icons';
 import { DelonACLModule } from '@delon/acl';
 import { DelonFormModule } from '@delon/form';
 import { AlainThemeModule } from '@delon/theme';
 import { NzBreadCrumbModule } from 'ng-zorro-antd/breadcrumb';
+import { NzButtonModule } from 'ng-zorro-antd/button';
+import { NzDividerComponent } from 'ng-zorro-antd/divider';
+import { NzIconModule } from 'ng-zorro-antd/icon';
+import { NzInputModule } from 'ng-zorro-antd/input';
+import { NzRadioComponent, NzRadioGroupComponent } from 'ng-zorro-antd/radio';
+import { NzSpinModule } from 'ng-zorro-antd/spin';
+import { NzSwitchComponent } from 'ng-zorro-antd/switch';
 import { NzTagModule } from 'ng-zorro-antd/tag';
 
-import { HelpMassageShowComponent } from './components/help-massage-show/help-massage-show.component';
-import { KeyValueInputComponent } from './components/key-value-input/key-value-input.component';
-import { MetricsFieldInputComponent } from './components/metrics-field-input/metrics-field-input.component';
+// Icon to be used for registration
+const icons: IconDefinition[] = [RobotOutline, CloseOutline, SendOutline];
+
+import { AiBotComponent } from './components/ai-bot/ai-bot.component';
+import { ConfigurableFieldComponent } from './components/configurable-field/configurable-field.component';
+import { FormFieldComponent } from './components/form-field/form-field.component';
+import { HelpMessageShowComponent } from './components/help-message-show/help-message-show.component';
+import { LabelSelectorComponent } from './components/label-selector/label-selector.component';
+import { MonitorSelectMenuComponent } from './components/monitor-select-menu/monitor-select-menu.component';
+import { MultiFuncInputComponent } from './components/multi-func-input/multi-func-input.component';
+import { ToolbarComponent } from './components/toolbar/toolbar.component';
 import { ElapsedTimePipe } from './pipe/elapsed-time.pipe';
 import { I18nElsePipe } from './pipe/i18n-else.pipe';
 import { TimezonePipe } from './pipe/timezone.pipe';
@@ -18,7 +35,16 @@ import { SHARED_DELON_MODULES } from './shared-delon.module';
 import { SHARED_ZORRO_MODULES } from './shared-zorro.module';
 
 const ThirdModules: Array<Type<void>> = [];
-const COMPONENTS: Array<Type<void>> = [KeyValueInputComponent, HelpMassageShowComponent, MetricsFieldInputComponent];
+const COMPONENTS: Array<Type<void>> = [
+  MultiFuncInputComponent,
+  HelpMessageShowComponent,
+  ToolbarComponent,
+  ConfigurableFieldComponent,
+  FormFieldComponent,
+  MonitorSelectMenuComponent,
+  AiBotComponent,
+  LabelSelectorComponent
+];
 const DIRECTIVES: Array<Type<void>> = [TimezonePipe, I18nElsePipe, ElapsedTimePipe];
 
 @NgModule({
@@ -34,9 +60,17 @@ const DIRECTIVES: Array<Type<void>> = [TimezonePipe, I18nElsePipe, ElapsedTimePi
     ...SHARED_ZORRO_MODULES,
     ...ThirdModules,
     NzBreadCrumbModule,
-    NzTagModule
+    NzTagModule,
+    NzDividerComponent,
+    NzRadioGroupComponent,
+    NzRadioComponent,
+    NzSwitchComponent,
+    NzButtonModule,
+    NzInputModule,
+    NzIconModule.forChild(icons),
+    NzSpinModule
   ],
-  declarations: [...COMPONENTS, ...DIRECTIVES, HelpMassageShowComponent],
+  declarations: [...COMPONENTS, ...DIRECTIVES, HelpMessageShowComponent],
   exports: [
     CommonModule,
     FormsModule,
